websim: microKORG2 4-voice polyphony (vox) + fix two latent x4-path bugs
Implements §D of the websim follow-up plan: vox now runs small (4-voice)
polyphony in the browser, opt-in via MK2_OSC_VOICES in mk2_osc_bridge.h
(default 1; waves/dummy stay mono). Adds a round-robin voice allocator
(noteOn/noteOff -> per-voice pitch + processor.voiceEvent) and an active-voice
downmix (averages held voices; idle voices excluded). The offline render
harness takes a comma-separated chord (WEBSIM_RENDER_ARGS="out.wav 57,60,64")
and check_render.py gained a Goertzel --peaks check; vox renders a stable
A-minor triad (all three tones present), while a single note fails the 3-peak
check (true discrimination).

Enabling vox's x4 oscillator path (never exercised by the shipped single-voice
scalar path) surfaced two real bugs, both fixed:

1. Buffer overflow (the cause of runaway divergence): vox's mOscBuffer is
   kMk2HalfVoices*kMk2BufferSize = 256 floats (the unit's hardware block is 64),
   but websim renders 128-sample Web Audio quanta. Mono writes mOscBuffer[i]
   (<=128, fits); the x4 path writes mOscBuffer[i*4+v] up to 511, overrunning
   256 and corrupting adjacent member arrays (mNoiseLevelMod -> spurious noise ->
   the high-Q formant filter rang up -> output diverged to ~1e3). Fix: the bridge
   renders in <= kMk2BufferSize-sample sub-blocks for poly (mono keeps its single
   128-sample call, unchanged).

2. Undefined behaviour in float_simd.h: the non-NEON si_i32x{2,4}qn_to_f32x{2,4}
   computed 1.f / (1 << qPoint); 1 << 31 is signed-int overflow. It was masked
   because UpdateEg only ever called it with a 0 phase. Fixed to
   1.f / (float)((uint64_t)1 << qPoint). ARM uses the NEON vcvtq_n_f32_s32
   branch and is unaffected (the #else branch never compiles on ARM).

Verified: build-all (27 units) and render-all stay green; mono units
(waves 440 Hz) unchanged.

+++ platform/microkorg2/common/utils/float_simd.h
@@ -1351,7 +1351,7 @@
  #if defined(NEON_SIMD_FP)
    return vcvtq_n_f32_s32(x, qPoint);
  #else
-   const float divisor = 1.f / (1 << qPoint);
+   const float divisor = 1.f / (float)((uint64_t)1 << qPoint);  // 1<<31 as signed int is UB
    const float32x4_t tmp = {{(float)x.val[0] * divisor, (float)x.val[1] * divisor, (float)x.val[2] * divisor, (float)x.val[3] * divisor}};
    return tmp;
  #endif
@@ -1363,7 +1363,7 @@
  #if defined(NEON_SIMD_FP)
    return vcvt_n_f32_s32(x, qPoint);
  #else
-   const float divisor = 1.f / (1 << qPoint);
+   const float divisor = 1.f / (float)((uint64_t)1 << qPoint);  // 1<<31 as signed int is UB
    const float32x2_t tmp = {{(float)x.val[0] * divisor, (float)x.val[1] * divisor}};
    return tmp;
  #endif
